Add service folder for greeting and quote services
