+++ src/services/greet/GreetingProvider.ts
@@ -1,0 +1,3 @@
+export default interface GreetingProvider {
+  getGreeting(): Promise<string>;
+}
