+++ src/services/quote/QuoteService.ts
@@ -1,0 +1,3 @@
+export default interface QuoteProvider {
+  getQuote(): Promise<{ author: string; text: string } | null>;
+}
